+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add refresh token to Dashboard
 + Add single fetch. Prepare for call detail
 + Add open close modal
++ Refactor actions
 
 ## 0.16.0
 + Add dashboard